# src/main/java/pl/tomaszko/s03e04/web/SearchRequest.java
package pl.tomaszko.s03e04.web;

import com.fasterxml.jackson.databind.JsonNode;
import jakarta.validation.constraints.NotNull;

public class SearchRequest {

    @NotNull
    private JsonNode params;

    public JsonNode getParams() {
        return params;
    }

    public void setParams(JsonNode params) {
        this.params = params;
    }
}